// src/types/resources/Patient.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { THumanName } from '../partials/HumanName';
import { TContactPoint } from '../partials/ContactPoint';
import { TDate } from '../simpleTypes/Date';
import { TDateTime } from '../simpleTypes/DateTime';
import { TAddress } from '../partials/Address';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TInt } from '../simpleTypes/Int';
import { TAttachment } from '../partials/Attachment';
import { TPatientContact } from '../partials/PatientContact';
import { TPatientCommunication } from '../partials/PatientCommunication';
import { TReference } from '../partials/Reference';
import { TPatientLink } from '../partials/PatientLink';

export type TPatient = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  active?: boolean;
  name?: THumanName[];
  telecom?: TContactPoint[];
  gender?: string;
  birthDate?: TDate;
  deceasedBoolean?: boolean;
  deceasedDateTime?: TDateTime;
  address?: TAddress[];
  maritalStatus?: TCodeableConcept;
  multipleBirthBoolean?: boolean;
  multipleBirthInteger?: TInt;
  photo?: TAttachment[];
  contact?: TPatientContact[];
  communication?: TPatientCommunication[];
  generalPractitioner?: TReference[];
  managingOrganization?: TReference;
  link?: TPatientLink[];
};